\name{createBushl}
\alias{createBushl}
\title{Create an orthogonal array using the Bush algorithm with alternate strength >= 3}
\usage{
  createBushl(q, ncol, lambda, bRandom)
}
\arguments{
  \item{q}{the number of symbols in the array}

  \item{ncol}{number of parameters or columns}

  \item{bRandom}{should the array be randomized}
}
\value{
  an orthogonal array
}
\description{
  Create an orthogonal array using the Bush algorithm with
  alternate strength >= 3
}

